--- Models/OperatingSystems.cs
@@ -1,8 +1,0 @@
-namespace OnionFinder.Models;
-
-public enum OperatingSystems
-{
-    Linux = 0,
-    Windows = 1,
-    MacOS = 3
-}
